Remove specific casters, consolidate to geometry cast
