+++ src/Commands/Utils/PostgisColumnInformation.php
@@ -2,7 +2,6 @@
 
 namespace Clickbar\Magellan\Commands\Utils;
 
-use Clickbar\Magellan\Cast\GeographyCast;
 use Clickbar\Magellan\Cast\GeometryCast;
 
 class PostgisColumnInformation
@@ -43,29 +42,16 @@ class PostgisColumnInformation
 
     public function getCasterClass(): string
     {
-        if ($this->geometry_type === 'geometry') {
-            return GeometryCast::class;
-        }
-
-        return GeographyCast::class;
+        return GeometryCast::class;
     }
 
     public function toCastValue(): string
     {
-        if ($this->geometry_type === 'geometry') {
-            return GeometryCast::class.":$this->srid";
-        }
-
-        return GeographyCast::class.":$this->srid";
-
+        return GeometryCast::class;
     }
 
     public function toCastLineCode(): string
     {
-        if ($this->geometry_type === 'geometry') {
-            return "'$this->column' => GeometryCast::class.':$this->srid',";
-        }
-
-        return "'$this->column' => GeographyCast::class.':$this->srid',";
+        return "'$this->column' => GeometryCast::class,";
     }
 }
